Fix error in the JSON initial population generation
The Json structures were not being added to the initial population since
there was a check that made sure the initial population never surpassed
a certain value. This commit fixes that.

It also only replaces the string statements with JSON statements in 50%
of the cases. This is done since a ststement can have multiple string
statements and not all of them make sense to be JSON.